Avoid copying the layer to the undo buffer at beginning of brushstroke
Instead we perform this work at the end of the brush stroke, or in
general after any operation which modifies the current layer. 

This allows us to remove that latency from the beginning of drawing
operations (e.g. brush stroke start) and push it to the end where it is
much less obtrusive.

This prevents a big lag spike at the beginning of brush strokes in
crappy browsers like IE.